Release v1.2.3 (#27)
- 不正な raw 指し手、行き所のない駒、強制成り、駒情報なし Move32 の適用を正しく検証し、王手状態の破損を防止
- SFEN、StaticBook、SAZPACK、SBINPACK、KIF、KI2、CSA、JKF、YaneuraOu 定跡、HCP、PACK の不正入力を安全に拒否し、HCP / PackedSfen の表現不能局面をエラーとして報告
- KI2 の曖昧な3枚駒局面で「左上」などの複合修飾を出力し、export から parse への round-trip を保証
- KIF/KI2 の全角括弧、秒読み、複数行コメント、分岐位置、駒落ちの上手・下手表記を修正
- 千日手裁定を探索用16手履歴窓から分離し、保持された全履歴から長周期の千日手を判定
- Record のマージを原子的にし、深い棋譜木の処理を反復実装へ変更するとともに、Python 棋譜パーサで GIL を解放し、default features の Rust 構成を CI 対象へ追加

+++ crates/rsshogi/src/records/formats/common.rs
@@ -3,7 +3,7 @@ use encoding_rs::SHIFT_JIS;
 use std::collections::HashMap;
 
 pub(crate) type BoardMap = HashMap<(u8, u8), (char, String)>;
-pub(crate) type HandCounts = HashMap<char, HashMap<String, u8>>;
+pub(crate) type HandCounts = HashMap<char, HashMap<String, u32>>;
 
 /// テキストエンコーディング。
 ///
